1.修改下单页面数据显示。 2.充电完成后订单展示页面数据修改。 3.充电控制页面数据和逻辑完善。 4.站点详情页面点击可以充电的桩跳转到下单页面 5.OrderBean实体类接受服务器的数据类型改善。 6.修改了调试的url

+++ app/src/main/java/com/optimumnano/quickcharge/bean/OrderBean.java
@@ -26,7 +26,7 @@ public class OrderBean implements Serializable{
     /** 充电桩类型 **/
     public String pile_type;
     /** 预支付金额 **/
-    public int frozen_cash;
+    public Double frozen_cash;
     /** 实际充电费用 **/
     public double charge_cash;
     /** 充电桩功率 **/
